avoid negative trap numbers by definition
git-svn-id: svn+ssh://market.scs.stanford.edu/home/u1/asbestos/repo/jos64/trunk@3629 927ebb28-5305-0410-b3ca-c4b640df69b7

+++ kern/kern/prof.h
@@ -6,7 +6,7 @@
 
 void prof_init(void) ;
 void prof_syscall(uint64_t num, uint64_t time) ;
-void prof_trap(int num, uint64_t time) ;
+void prof_trap(uint64_t num, uint64_t time) ;
 void prof_user(uint64_t time) ;
 void prof_thread(uint64_t tid, uint64_t time);
 void prof_print(void) ;
